requirements.txt
torch+cu11.6
torchvision
opencv-python
pillow
tqdm
scikit-learn  
scipy
numpy
onnx
onnxruntime-gpu
timm
torchinfo
tensorboardX
mnn # below python3.8